assigned footprints, fixed electrical rules checker errors
I fixed the ERC errors by adding no connect flags to the unused pins, I assigned the footprints for all the components, and initialized the PCB with all the components. I also forgot to add the mounting holes and mousebites to the last commit, so I added those as well. Next commit will be routing all the components, then full CAD, then firmware (looking to use KMK as that's what I used for the Macropad)