finished routing, added the back Silkscreen, fixed DRC
I finished the leftover routing, added some images to the back silkscreen, and went through the DRC errors. There are still some problems with the USB-C port's clearance, but otherwise, the PCB is almost done. I'll try to get the Cad and FW done this week so I can wrap up and submit on Sat. PCB way says ~54$ for the pcb.